Adjust window size and create custom render script
The new render script just changes the camera to fixed-fit projection. I see that with Defold I will have to be careful to ensure the game view is consistent when the window is resized or on different screen sizes (think about multiplayer information) because by default the camera will show more or less information based on window size.