Set up basic enemy and implement basic movement
The enemy follows the player forever and it is tightly coupled to exactly one player. The player also posts a message to the exactly one enemy about its position. This needs to be architected for many enemies to many players, and I will learn to do that in time with the message passing system and after implementing enemy factories.