Attack closest nearby enemy
I think. It seems basically correct though sometimes it doesn't look right. I think that's just because of the cooldown on the sword. Will correct later if I need to. I'm sorting the table every frame for now because doing it only when enemies enter or leave is not good enough to keep them sorted.