Implement experience vial pickup
The player collides with the pickup, it posts a message to the player and deletes itself, and the player increments its experience value.

`gain_experience()` exists, similarly to `take_damage()`, for inevitable future functionality that coincides with incrementing the experience value.

I must say that message passing is a very natural fit for me. It just works.