Only handle abilities if an enemy is nearby
Now the sword ability does not spawn on the player and abilities are saved until an enemy is nearby. In the future, certain abilities will probably go off regardless (non-targeted abilities). I'm also wondering if I should be sorting the nearby enemies once a frame or if I should return out of this function if none of the ability cooldowns are going off.